+++ tests/plugin/index.test.ts
@@ -1,66 +1,11 @@
 import fs from 'fs';
 import path from 'path';
-import { js_beautify } from 'js-beautify';
 
-import HandlebarsTransformer from '../lib/handlebars-transformer';
+import { HandlebarsPluginOptions } from '../../lib/types/plugin-options';
+import { lookupHelperRegistration, lookupPartialRegistration } from './utils/utils';
+import { removeOutputDir, testTemplate } from './utils';
 
-import { HandlebarsPluginOptions } from '../lib/types/plugin-options';
-import { CompileResult } from '../lib/types/handlebars';
-
-type TestFn = (err: Error, output: CompileResult) => {}
-
-const testFunctionsDir = '../runtime-tests/functions'
-const absoluteTestFunctionsDir = path.resolve(__dirname, testFunctionsDir)
-if (fs.existsSync(absoluteTestFunctionsDir)) {
-    fs.rmdirSync(absoluteTestFunctionsDir, { recursive: true });
-}
-
-function loadTemplate(templatePath: string) {
-    return fs.readFileSync(templatePath).toString();
-}
-
-function createFunctionFile(template: string, output: CompileResult) {
-    const templatePath = path.join(__dirname, template)
-    const extname = path.extname(template)
-    const outputTemplate = path.join(__dirname, testFunctionsDir, template)
-    const outputDir = path.dirname(outputTemplate)
-    const outputName = `${path.basename(templatePath, extname)}.js`
-    const outputPath = path.join(outputDir, outputName)
-    if (!fs.existsSync(outputDir)) {
-        fs.mkdirSync(outputDir, { recursive: true });
-    }
-    let code: string = output?.code || ''
-    code = !code ? code : js_beautify(code)
-    fs.writeFileSync(outputPath, code)
-}
-
-function testTemplate(template: string, pluginOptions: HandlebarsPluginOptions, testFn: TestFn) {
-
-    let err: Error | undefined
-    const hbsTransformer = new HandlebarsTransformer(pluginOptions)
-
-    const templatePath = path.join(__dirname, template)
-    
-    let source: string | null = null
-    try {
-        source = loadTemplate(templatePath)
-    } catch (e) {
-        err = e
-        return
-    }
-
-    let output: CompileResult | null = null
-    try {
-        output = hbsTransformer.transform(source, templatePath)
-    } catch (e) {
-        err = e
-    }
-
-    testFn(err, output)
-
-    createFunctionFile(template, output)
-
-}
+removeOutputDir()
 
 describe('Handlebars Transformer', () => {
 
@@ -85,7 +30,7 @@ describe('Handlebars Transformer', () => {
             './invalid-syntax-error.hbs',
             pluginOptions,
             async (err, output) => {
-                const catchError = err.message.indexOf("Parse error") >= 0
+                const catchError = err?.message.includes("Parse error")
                 expect(catchError).toBe(true)
             }
         )
@@ -101,7 +46,7 @@ describe('Handlebars Transformer', () => {
             './invalid-unknown-helpers.hbs',
             pluginOptions,
             async (err, output) => {
-                const catchError = err.message.indexOf("You specified knownHelpersOnly") >= 0
+                const catchError = err?.message.includes("You specified knownHelpersOnly")
                 expect(catchError).toBe(true)
             }
         )
@@ -146,7 +91,47 @@ describe('Handlebars Transformer', () => {
             './with-helpers-commonjs.hbs',
             pluginOptions,
             async (err, output) => {
-                const catchOutput = output.code.indexOf("Handlebars.registerHelper('description") >= 0
+                const catchOutput = lookupHelperRegistration('descriptionHelper', output?.code)
+                expect(catchOutput).toBe(true)
+            }
+        )
+    })
+
+    it('should be able to use block helpers', () => {
+
+        const pluginOptions: HandlebarsPluginOptions = {
+            helpers: {
+                list: function(items: Array<object>, options) {
+                    const itemsAsHtml = items.map(item => "<li>" + options.fn(item) + "</li>");
+                    return "<ul>\n" + itemsAsHtml.join("\n") + "\n</ul>";
+                }
+            }
+        }
+
+        testTemplate(
+            './with-block-helpers.hbs',
+            pluginOptions,
+            async (err, output) => {
+                const catchOutput = lookupHelperRegistration('list', output.code)
+                expect(catchOutput).toBe(true)
+            }
+        )
+
+    })
+
+    it('should allow partials to be passed through the plugin options', async () => {
+
+        const pluginOptions: HandlebarsPluginOptions = {
+            partials: {
+                otherPartial: fs.readFileSync(path.join(__dirname, './partialDirs/anotherDir/otherPartial.hbs')).toString()
+            }
+        }
+
+        testTemplate(
+            './with-plugin-partial.hbs',
+            pluginOptions,
+            async (err, output) => {
+                const catchOutput = lookupPartialRegistration('otherPartial', output?.code)
                 expect(catchOutput).toBe(true)
             }
         )
@@ -178,8 +163,8 @@ describe('Handlebars Transformer', () => {
             './with-dir-partials.hbs',
             pluginOptions,
             async (err, output) => {
-                const catchOutput1 = output?.code.indexOf("Handlebars.registerPartial('partialDirs/otherPartial") >= 0
-                const catchOutput2 = output?.code.indexOf("Handlebars.registerPartial('partialDirs/anotherDir/otherPartial") >= 0
+                const catchOutput1 = lookupPartialRegistration('partialDirs/otherPartial', output?.code)
+                const catchOutput2 = lookupPartialRegistration('partialDirs/anotherDir/otherPartial', output?.code)
                 expect(catchOutput1 && catchOutput2).toBe(true)
             }
         )
@@ -187,6 +172,48 @@ describe('Handlebars Transformer', () => {
 
     it.todo('should allow nested partials')
     it.todo('should allow resolving parent paths in partials')
+
+    it('should allow partials from ancestor directory', async () => {
+
+        const pluginOptions: HandlebarsPluginOptions = {}
+
+        testTemplate(
+            './nested-templates/nested/with-ancestor-dir-partial.hbs',
+            pluginOptions,
+            async (err, output) => {
+                const catchOutput = lookupPartialRegistration('../../some-partial', output?.code)
+                expect(catchOutput).toBe(true)
+            }
+        )
+    })
+
+    it('should allow partials from parent directory', async () => {
+
+        const pluginOptions: HandlebarsPluginOptions = {}
+
+        testTemplate(
+            './nested-templates/with-parent-dir-partial.hbs',
+            pluginOptions,
+            async (err, output) => {
+                const catchOutput = lookupPartialRegistration('../some-partial', output?.code)
+                expect(catchOutput).toBe(true)
+            }
+        )
+    })
+
+    it('should allow partials from cousin directory', async () => {
+
+        const pluginOptions: HandlebarsPluginOptions = {}
+
+        testTemplate(
+            './nested-templates/with-cousin-dir-partial.hbs',
+            pluginOptions,
+            async (err, output) => {
+                const catchOutput = lookupPartialRegistration('../partialDirs/anotherDir/otherPartial', output?.code)
+                expect(catchOutput).toBe(true)
+            }
+        )
+    })
 
     it('should use failover content of the partial block if it refers to non-existent partial', async () => {
 
@@ -196,7 +223,7 @@ describe('Handlebars Transformer', () => {
             './with-partial-block.hbs',
             pluginOptions,
             async (err, output) => {
-                const catchOutput = output.code.indexOf("<div>Failover</div>") >= 0
+                const catchOutput = output?.code.indexOf("<div>Failover</div>") >= 0
                 expect(catchOutput).toBe(true)
             }
         )
@@ -210,7 +237,7 @@ describe('Handlebars Transformer', () => {
             './with-inline-partial.hbs',
             pluginOptions,
             async (err, output) => {
-                const catchOutput = output.code.indexOf("printFoo") >= 0
+                const catchOutput = output?.code.indexOf("printFoo") >= 0
                 expect(catchOutput).toBe(true)
             }
         )